Remove extraction over node from selection from the init.py To a separate file
